saga_architect/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
pydantic
chromadb
sentence-transformers
python-frontmatter